requirements.txt
PyAudio==0.2.14
websockets==13.1
pre-commit==4.0.1
python-dotenv==1.0.1
pyyaml==6.0.2
sounddevice==0.5.0
numpy==2.1.2
scipy==1.14.1
